// backend/shared/utils/logger.d.ts
export declare const logger: {
    info: (message: string, ...args: any[]) => void;
    warn: (message: string, ...args: any[]) => void;
    error: (message: string, ...args: any[]) => void;
    debug: (message: string, ...args: any[]) => void;
};
export default logger;
//# sourceMappingURL=logger.d.ts.map